Se ordenan archivos y carpetas del proyecto
