+++ src/app.js
@@ -1,10 +1,9 @@
 import express from "express";
 import router from "./routes/index.js";
-import handlebars from "express-handlebars";
+import hbs from "./configs/handlebars.js";
 import morgan from "morgan";
-import { isEmptyArray } from "./utils.js";
-import { mongodb } from "./database.js";
-import socketioHandler from "./socket.js";
+import { mongodb } from "./services/mongo.js";
+import socketioHandler from "./helpers/socket.js";
 
 const app = express();
 
@@ -17,16 +16,6 @@ app.use(express.urlencoded({ extended: true }));
 app.use(morgan("dev"));
 
 //Navegador
-const hbs = handlebars.create({
-  runtimeOptions: {
-    allowProtoPropertiesByDefault: true,
-    allowProtoMethodsByDefault: true,
-  },
-  helpers: {
-    isEmptyArray: isEmptyArray,
-  },
-});
-
 app.use("/static", express.static("./src/public"));
 app.engine("handlebars", hbs.engine);
 app.set("view engine", "handlebars");
@@ -35,7 +24,7 @@ app.set("views", "./src/views");
 router(app);
 
 app.get("/", (req, res) => {
-  res.render("index", {title: "Atlas Tech"});
+  res.render("index", { title: "Atlas Tech" });
 });
 
 const httpServer = app.listen(PORT, () => {
